requirements.txt
Flask==3.0.0
Flask-Caching==2.1.0
Flask-Cors==4.0.0
flask-session==0.6.0
google-generativeai==0.8.6
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
gunicorn==21.2.0
werkzeug==3.0.2